// src/app/interface/interface.ts
export interface HomeworldInterface {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}

export interface StarshipsInterface {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  consumables: string;
  hyperdrive_rating: string;
  MGLT: string;
  starship_class: string;
  pilots: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}

export interface VehiclesInterface {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  consumables: string;
  vehicle_class: string;
  pilots: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}

export interface FilmsInterface {
  characters: string[];
  created: string;
  director: string;
  edited: string;
  episode_id: number;
  opening_crawl: string;
  planets: string[];
  producer: string;
  release_date: string;
  species: string[];
  starships: string[];
  title: string;
  url: string;
  vehicles: string[];
}

export interface CharacterInterface {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  filmsData: FilmsInterface[];
  species?: string[];
  vehicles: string[];
  vehiclesData: VehiclesInterface[];
  starships: string[];
  starshipsData: StarshipsInterface[];
  created: string;
  edited: string;
  url: string;
}

export interface DataFromApiInterface {
  count: number;
  next: string | null;
  previous: string | null;
  results: CharacterInterface[];
}

export interface InitialInterface {
  favoriteList: CharacterInterface[];
  dataFromApi: CharacterInterface[];
  isDarkTheme: boolean;
  isLoading: boolean;
  isDisplay: boolean;
  charactersAmount: number;
  isError: boolean;
  characterDetails: CharacterInterface[];
  pageNumber: number;
}